nVd Computations

+++ PicesLibraries/PicesLibrary/DataBase.cpp
@@ -4670,7 +4670,7 @@ void  DataBase::ImagesSizeDistributionByDepth (const KKStr&               cruise
     {
       if (!downCast)
       {
-        downCast = new ImageSizeDistribution (depthBinSize, initialValue, growthRate, endValue, startValues, endValues, log);
+        downCast = new ImageSizeDistribution ((float)depthBinSize, (float)initialValue, (float)growthRate, endValue, startValues, endValues, log);
       }
       sizeDistribution = downCast;
       if  (volSampledDownCast)
